Add comment POST and GET requests

+++ chat/src/main/java/com/chat/chat/MessegeRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 import com.chat.chat.Messege;
 
-public interface MessegeRepository extends CrudRepository<Messege, Integer>{
+public interface MessegeRepository extends CrudRepository<Messege, Long>{
 	List<Messege> findByServiceIdAndCustomerId(Long serviceId, Long customerId);
 }
